docs(architecture): [Antigravity] Define configuration ownership boundaries for plugin marketplace MVP

+++ Apps/ga-server/GA.BSP.Service/Services/MongoDbService.cs
@@ -194,8 +194,6 @@ public class MongoDbService
         }
     }
 
-    // TODO: Re-enable when ChordStatistics type is available in GA.Business.Orchestration
-    /*
     public async Task<ChordStatistics> GetChordStatisticsAsync()
     {
         var totalCount = await GetTotalChordCountAsync();
@@ -262,14 +260,13 @@ public class MongoDbService
 
         return new ChordStatistics
         {
-            TotalChords = totalCount,
+            TotalChords = (int)totalCount,
             QualityDistribution = qualityDistribution,
             ExtensionDistribution = extensionDistribution,
             StackingTypeDistribution = stackingDistribution,
             NoteCountDistribution = noteCountDistribution
         };
     }
-    */
 
     public async Task<List<Chord>> SearchChordsAsync(string query, int limit = 100)
     {
